pyproject.toml
# This project uses uv for dependency management
# Install: curl -LsSf https://astral.sh/uv/install.sh | sh
# Setup: uv sync
# Run: uv run python audit.py

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dw-auditor"
version = "1.0.0"
description = "High-performance data quality auditing for BigQuery & Snowflake with automatic relationship detection"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Valentin Cathelain", email = "valentin@example.com"}
]
maintainers = [
    {name = "Valentin Cathelain", email = "valentin@example.com"}
]
requires-python = ">=3.10"
keywords = [
    "data-quality",
    "data-warehouse",
    "bigquery",
    "snowflake",
    "data-profiling",
    "data-validation",
    "etl",
    "data-engineering"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
    "Typing :: Typed"
]

# Core dependencies
dependencies = [
    "polars>=0.20.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "ibis-framework[duckdb]>=9.0.0",
    "ibis-framework[bigquery]>=9.0.0",
    "ibis-framework[snowflake]>=9.0.0",
    "google-cloud-bigquery>=3.0.0",
    "snowflake-connector-python>=3.0.0",
]

[project.optional-dependencies]
# Development dependencies
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

# Documentation
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
]

# Enhanced CLI experience
rich = [
    "rich>=13.0.0",
    "tqdm>=4.65.0",
]

# All optional dependencies
all = [
    "dw-auditor[dev,docs,rich]",
]

[project.urls]
Homepage = "https://github.com/yourusername/database_audit"
Documentation = "https://github.com/yourusername/database_audit/blob/main/README.md"
Repository = "https://github.com/yourusername/database_audit"
"Bug Tracker" = "https://github.com/yourusername/database_audit/issues"
Changelog = "https://github.com/yourusername/database_audit/blob/main/CHANGELOG.md"

[project.scripts]
dw-audit = "audit:main"

# Tool configurations
# ============================================================================

[tool.setuptools]
packages = ["dw_auditor"]

[tool.setuptools.package-data]
dw_auditor = ["py.typed"]

# Black - Code formatter
[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
  | \.venv
  | audit_env
  | audit_results
  | __pycache__
  | \.pytest_cache
)/
'''

# Ruff - Fast Python linter
[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "ARG",  # flake8-unused-arguments
    "SIM",  # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # function calls in argument defaults
    "B905",  # zip strict argument
    "C901",  # complexity checks
]
exclude = [
    ".git",
    ".venv",
    "audit_env",
    "audit_results",
    "__pycache__",
    ".pytest_cache",
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

[tool.ruff.isort]
known-first-party = ["dw_auditor"]

# MyPy - Static type checker
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Set to true for stricter typing
ignore_missing_imports = true
exclude = [
    "audit_env",
    "audit_results",
    ".venv",
]

# Pytest - Testing framework
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--strict-markers",
    "--cov=dw_auditor",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "bigquery: marks tests requiring BigQuery access",
    "snowflake: marks tests requiring Snowflake access",
]

# Coverage
[tool.coverage.run]
source = ["dw_auditor"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/audit_env/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
